Move the seek bar to a separate component
